Revert "feat: sdk tween sequence component support (sdk migration) (#5790)"
This reverts commit c20b87eae74cce0092bac8406ea7dca6b3a42f7b.
