+++ Explorer/Assets/DCL/Infrastructure/ProtobufPartialClasses/IDirtyMarker.cs
@@ -117,11 +117,6 @@ namespace DCL.ECSComponents
         public bool IsDirty { get; set; }
     }
 
-    public partial class PBTweenSequence : IDirtyMarker
-    {
-        public bool IsDirty { get; set; }
-    }
-
     public partial class PBMapPin : IDirtyMarker
     {
         public bool IsDirty { get; set; }
