+++ Explorer/Assets/DCL/Infrastructure/SceneRunner/Scene/SceneData.cs
@@ -5,7 +5,6 @@ using SceneRuntime.ScenePermissions;
 using DCL.SceneRunner.Scene;
 using System;
 using System.Collections.Generic;
-using System.Text.RegularExpressions;
 using UnityEngine;
 using Utility;
 
@@ -46,7 +45,7 @@ namespace SceneRunner.Scene
             SceneEntityDefinition = sceneDefinition;
             StaticSceneMessages = staticSceneMessages;
             Parcels = parcels;
-            SceneShortInfo = new SceneShortInfo(baseParcel, sceneDefinition.id, sceneDefinition.metadata.sdkVersion);
+            SceneShortInfo = new SceneShortInfo(baseParcel, sceneDefinition.id);
             Geometry = geometry;
             InitialSceneStateInfo = initialSceneStateInfo;
         }
@@ -137,82 +136,5 @@ namespace SceneRunner.Scene
         public bool IsSdk7() =>
             SceneEntityDefinition.metadata.runtimeVersion == "7";
 
-        /// <summary>
-        ///     Gets the specific SDK version (e.g.: "7.5.6" or "7.11.1-18575373673.commit-f7e8f37")
-        ///     that is injected in Scenes Metadata during their deployment
-        /// </summary>
-        public string GetSDKVersion() =>
-            SceneEntityDefinition.metadata.sdkVersion;
-
-        /// <summary>
-        ///     Checks if the scene's SDK version is the specified version or higher
-        /// </summary>
-        /// <param name="minVersion">Minimum version to check (e.g., "7.5.0")</param>
-        /// <returns>True if scene SDK version >= minVersion, false if version unknown or less than minVersion</returns>
-        public bool IsSDKVersionOrHigher(string minVersion)
-        {
-            string sdkVersion = SceneEntityDefinition.metadata.sdkVersion;
-
-            if (string.IsNullOrEmpty(sdkVersion) || string.IsNullOrEmpty(minVersion))
-                return false;
-
-            try
-            {
-                Version sceneVersion = ParseSDKVersion(sdkVersion);
-                Version minimumVersion = ParseSDKVersion(minVersion);
-
-                return sceneVersion >= minimumVersion;
-            }
-            catch
-            {
-                // If version parsing fails, return false for safety
-                return false;
-            }
-        }
-
-        /// <summary>
-        ///     Checks if the scene's SDK version matches a specific version
-        /// </summary>
-        public bool IsSDKVersion(string version)
-        {
-            string sdkVersion = SceneEntityDefinition.metadata.sdkVersion;
-
-            if (string.IsNullOrEmpty(sdkVersion) || string.IsNullOrEmpty(version))
-                return false;
-
-            try
-            {
-                Version sceneVersion = ParseSDKVersion(sdkVersion);
-                Version targetVersion = ParseSDKVersion(version);
-
-                return sceneVersion == targetVersion;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
-        private static Version ParseSDKVersion(string versionString)
-        {
-            // Handle versions like "7.5.6", "7.5", "7", or "7.11.1-18575373673.commit-f7e8f37"
-            // Extract only the semantic version part (major.minor.patch) before any pre-release or build metadata
-
-            // Use regex to extract version numbers at the start of the string
-            // Pattern: one or more groups of digits separated by dots
-            Match match = Regex.Match(versionString, @"^(\d+)(?:\.(\d+))?(?:\.(\d+))?");
-
-            if (!match.Success)
-                throw new FormatException($"Invalid version format: {versionString}");
-
-            string major = match.Groups[1].Value;
-            string minor = match.Groups[2].Success ? match.Groups[2].Value : "0";
-            string patch = match.Groups[3].Success ? match.Groups[3].Value : "0";
-
-            // Construct a clean version string
-            string cleanVersion = $"{major}.{minor}.{patch}";
-
-            return new Version(cleanVersion);
-        }
     }
 }
