+++ Explorer/Assets/DCL/Ipfs/SceneMetadata.cs
@@ -14,7 +14,6 @@ namespace DCL.Ipfs
         public string main;
         public SceneMetadataScene scene;
         public string runtimeVersion;
-        public string sdkVersion;
         public List<string> allowedMediaHostnames;
         public List<string> requiredPermissions;
         public List<SpawnPoint>? spawnPoints;
