+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/Diagnostics/SceneShortInfo.cs
@@ -10,25 +10,24 @@ namespace DCL.Diagnostics
     {
         public readonly Vector2Int BaseParcel;
         public readonly string Name;
-        public readonly string? SdkVersion;
 
-        public SceneShortInfo(Vector2Int baseParcel, string name, string? sdkVersion = null)
+
+        public SceneShortInfo(Vector2Int baseParcel, string name)
         {
             BaseParcel = baseParcel;
             Name = name;
-            SdkVersion = sdkVersion;
         }
 
         public override string ToString() =>
             $"({BaseParcel.x},{BaseParcel.y}) - {Name}";
 
         public bool Equals(SceneShortInfo other) =>
-            BaseParcel.Equals(other.BaseParcel) && Name == other.Name && SdkVersion == other.SdkVersion;
+            BaseParcel.Equals(other.BaseParcel) && Name == other.Name;
 
         public override bool Equals(object? obj) =>
             obj is SceneShortInfo other && Equals(other);
 
         public override int GetHashCode() =>
-            HashCode.Combine(BaseParcel, Name, SdkVersion);
+            HashCode.Combine(BaseParcel, Name);
     }
 }
