+++ Explorer/Assets/DCL/PluginSystem/World/TweenPlugin.cs
@@ -8,18 +8,11 @@ using DCL.SDKComponents.Tween.Components;
 using TweenCleanUpSystem = DCL.SDKComponents.Tween.TweenCleanUpSystem;
 using TweenLoaderSystem = DCL.SDKComponents.Tween.TweenLoaderSystem;
 using TweenUpdaterSystem = DCL.SDKComponents.Tween.TweenUpdaterSystem;
-using TweenSequenceLoaderSystem = DCL.SDKComponents.Tween.TweenSequenceLoaderSystem;
-using TweenSequenceUpdaterSystem = DCL.SDKComponents.Tween.TweenSequenceUpdaterSystem;
 
 namespace DCL.PluginSystem.World
 {
     public class TweenPlugin : IDCLWorldPluginWithoutSettings
     {
-        // Previous SDK versions have the TweenSequence logic running in the SDK Runtime,
-        // we need to still support those already deployed scenes and we
-        // cannot have both running at the same time (SDK Runtime & Explorer)
-        private const string MIN_TWEEN_SEQUENCE_SDK_VERSION = "7.13.0";
-
         private readonly TweenerPool tweenerPool;
 
         public TweenPlugin()
@@ -29,23 +22,12 @@ namespace DCL.PluginSystem.World
 
         public void InjectToWorld(ref ArchSystemsWorldBuilder<Arch.Core.World> builder, in ECSWorldInstanceSharedDependencies sharedDependencies, in PersistentEntities persistentEntities, List<IFinalizeWorldSystem> finalizeWorldSystems, List<ISceneIsCurrentListener> sceneIsCurrentListeners)
         {
-            bool nativeTweenSequenceSupport = sharedDependencies.SceneData.IsSDKVersionOrHigher(MIN_TWEEN_SEQUENCE_SDK_VERSION);
-
             ResetDirtyFlagSystem<PBTween>.InjectToWorld(ref builder);
+            TweenLoaderSystem.InjectToWorld(ref builder);
 
-            if (nativeTweenSequenceSupport)
-            {
-                ResetDirtyFlagSystem<PBTweenSequence>.InjectToWorld(ref builder);
-                TweenSequenceLoaderSystem.InjectToWorld(ref builder);
-                TweenSequenceUpdaterSystem.InjectToWorld(ref builder, sharedDependencies.EcsToCRDTWriter, tweenerPool, sharedDependencies.SceneStateProvider);
-            }
-            else
-            {
-                TweenLoaderSystem.InjectToWorld(ref builder);
-                TweenUpdaterSystem.InjectToWorld(ref builder, sharedDependencies.EcsToCRDTWriter, tweenerPool, sharedDependencies.SceneStateProvider);
-            }
+            TweenUpdaterSystem.InjectToWorld(ref builder, sharedDependencies.EcsToCRDTWriter, tweenerPool, sharedDependencies.SceneStateProvider);
 
-            finalizeWorldSystems.Add(TweenCleanUpSystem.InjectToWorld(ref builder, sharedDependencies.EcsToCRDTWriter, tweenerPool, nativeTweenSequenceSupport));
+            finalizeWorldSystems.Add(TweenCleanUpSystem.InjectToWorld(ref builder, sharedDependencies.EcsToCRDTWriter, tweenerPool));
         }
     }
 }
