+++ Explorer/Assets/DCL/SDKComponents/Tween/Components/CustomTweener/ICustomTweener.cs
@@ -12,7 +12,7 @@ namespace DCL.SDKComponents.Tween.Components
 
         void Rewind();
 
-        void Kill(bool complete);
+            void Kill(bool complete);
 
         bool IsPaused();
