--- Explorer/Assets/DCL/SDKComponents/Tween/Components/CustomTweener/SequenceTweener.cs.meta
@@ -1,16 +1,0 @@
-fileFormatVersion: 2
-guid: d1c6f4b8a9e5479b0d3f4e6f7b8c9d0e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-
-
-
-
-
