--- Explorer/Assets/DCL/SDKComponents/Tween/Components/SDKTweenSequenceComponent.cs.meta
@@ -1,16 +1,0 @@
-fileFormatVersion: 2
-guid: b8a4d2f9e7c3469a9b1e2d4c5f6a7b8c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-
-
-
-
-
