--- Explorer/Assets/DCL/SDKComponents/Tween/Systems/TweenSequenceLoaderSystem.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bfe084ef20a7d45119a295e6b9f1009e
