--- Explorer/Assets/DCL/SDKComponents/Tween/Systems/TweenSequenceUpdaterSystem.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a9b8c7d6e5f4469a0d3e2f1d5a6b7c8e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-
