--- Explorer/Assets/DCL/SDKComponents/Tween/Tests/TweenSequenceLoaderSystemShould.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 45239728fa82649d795dbbeac3dd25c8
