--- Explorer/Assets/DCL/SDKComponents/Tween/Tests/TweenSequenceSystemShould.cs.meta
@@ -1,16 +1,0 @@
-fileFormatVersion: 2
-guid: e3a7f8d4c9b5469a0d3e4f5d6a7b8c9e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-
-
-
-
-
